+++ docs/src/components/HomeScreen.js
@@ -5,9 +5,10 @@ import firebase from 'firebase';
 import { Link } from 'react-router-dom';
 
 const Container = styled.div`
-  padding: 50px 15%;
-  min-height: 100vh;
-  background: #eee;
+  padding: 50px 15% 0px 15%;
+  background: #fff;
+  color: #424242;
+  height: 100vh;
 `;
 
 const HomeScreen = props => {
@@ -15,33 +16,33 @@ const HomeScreen = props => {
   return (
     <Container>
       <h2>And that's how easy it is to get started with Frothy!</h2>
-      <h4>
-        Well, there's a little Firebase configuration that you will also need to
-        do but it's a painless process.
-      </h4>
+      <p>
+        Well, there's a little Firebase configuration, but it's a painless
+        process.
+      </p>
 
-      <h4>
+      <h2>About this demo</h2>
+
+      <p>
         This demo is using the Firebase <code>onAuthStateChanged</code> listener
         along with React-Router to control route navigation based on the
         authentication state.
-      </h4>
-      <h4>
+      </p>
+      <p>
         If you try to navigate back to the login screen, you should be
-        redirected back to this /home route.
-      </h4>
-      <h4>
-        Conversely, if you try to navigate to this /home route while not logged
-        in, you should be redirected back to the login route.
-      </h4>
+        redirected back to this /home route. Conversely, if you try to navigate
+        to this /home route while not logged in, you should be redirected back
+        to the login route.
+      </p>
 
-      <Link to="/">
+      <Link to="/" style={{ textDecoration: 'none' }}>
         <Button variant="flat" color="primary">
           Try to navigate to login route
         </Button>
       </Link>
+
       <br />
-      <br />
-      <h3>
+      <p>
         Learn how to get started using Frothy in your own app on{' '}
         <a
           href="https://github.com/gojutin/frothy#readme"
@@ -51,7 +52,7 @@ const HomeScreen = props => {
           {' '}
           GitHub
         </a>.
-      </h3>
+      </p>
       <br />
       <Button variant="raised" color="secondary" onClick={logout}>
         LOGOUT