+++ docs/src/components/options/CodeBox.js
@@ -1,11 +1,25 @@
 import React from 'react';
-
+import styled from 'styled-components';
 import SyntaxHighlighter from 'react-syntax-highlighter';
 import { docco } from 'react-syntax-highlighter/styles/hljs';
 
-const CodeBox = props => (
-  <div style={{ textAlign: 'left', width: '90%', margin: '0 auto' }}>
-    <SyntaxHighlighter language="javascript" style={docco}>{`
+const Container = styled.div`
+  text-align: left;
+  width: 90%;
+  max-width: 300px;
+  margin: 0 auto;
+`;
+
+const CodeBox = props => {
+  return (
+    <Container>
+      <SyntaxHighlighter language="javascript" style={docco}>{`
+// This code snippet contains the
+// default props, which do not
+// need to be explicitly defined.
+
+import Frothy from 'frothy';
+
 <Frothy
   modal={${props.modal}}
   themeColor='${props.themeColor}'
@@ -19,9 +33,11 @@ const CodeBox = props => (
   emailLogin={${props.emailLogin}}
   emailRemember={${props.emailRemember}}
   emailSignup={${props.emailSignup}}
+  agree={${props.agree}}
   passwordReset={${props.passwordReset}}
   phone={${props.phone}}
   recaptcha='${props.recaptcha}'
+  recaptchaBadge='${props.recaptchaBadge}'
   anonymous={${props.anonymous}}
   google={${props.google}}
   facebook={${props.facebook}}
@@ -29,7 +45,8 @@ const CodeBox = props => (
   github={${props.github}}
 />
 `}</SyntaxHighlighter>
-  </div>
-);
+    </Container>
+  );
+};
 
 export default CodeBox;