+++ docs/src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 
 import './index.css';
 import App from './App';
+import registerServiceWorker from './registerServiceWorker';
 
 import firebase from 'firebase/app';
 
@@ -18,3 +19,4 @@ const config = {
 firebase.initializeApp(config);
 
 ReactDOM.render(<App />, document.getElementById('root'));
+registerServiceWorker();