+++ src/components/Header.js
@@ -1,21 +1,17 @@
 import React from 'react';
+import styled from 'styled-components';
 
-const Header = props => {
-  const Title = props.title;
-  const styles = {
-    wrapper: {
-      textAlign: 'center',
-      padding: '10px',
-      borderRadius: props.rounded,
-      color: props.titleColor,
-      backgroundColor: props.titleBackgroundColor,
-    },
-  };
-  return (
-    <div style={styles.wrapper}>
-      {typeof Title === 'string' ? <h3>{Title}</h3> : Title}
-    </div>
-  );
-};
+const StyledHeader = styled.div`
+  text-align: center;
+  padding: 10px;
+  color: ${props => props.titleColor};
+  background-color: ${props => props.titleBackgroundColor};
+`;
+
+const Header = ({ title: Title, ...rest }) => (
+  <StyledHeader {...rest}>
+    {typeof Title === 'string' ? <h3>{Title}</h3> : Title}
+  </StyledHeader>
+);
 
 export default Header;