+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.9] - 2018-05-13
+
+Happy Mother's Day to all the moms out there!
+
+### Changed
+
+* Updated various styles including a lighter background color for the anonymous login button.
+* Replaced floating Recaptcha badge with inline badge in phone login form.
+* Fixed styling issue where rounded edges and shadows were not displaying correctly as a modal.
+
+### Added
+
+* Added `agree` and `agreeMessage` props, which generate an "Agree to Terms" checkbox in the email signup form.
+* Added `modalOverlay` prop to toggle the modal overlay on or off.
+* Added `recaptchaBadge` prop to control position of the required Recaptcha badge.
+* Added animations for modal mount and unmount.
+
+### Removed
+
+* Removed Recaptcha for email signup. While Recaptcha support is enabled by default for the phone login, it appears that an API key may be required to enable Recaptcha on email signup. This is being removed until it can be properly addressed.
+
 ## [0.0.8] - 2018-05-12
 
 ### Changed